refactor(conversation): replace modal with inline health check card
- Remove AgentHealthCheckModal component in favor of inline card
- Display health check directly in conversation messages
- Add fixed height scrolling for multiple agents (max-h-120px for checking, max-h-160px for results)
- Improve UX with smaller, less intrusive card design
- Keep all health check and auto-switching functionality
